[DONE] cow authorization

+++ src/app/modules/cow/cow.service.ts
@@ -5,6 +5,12 @@ import { cowSearchableFields } from './cow.constants';
 import { ICow, ICowFilter } from './cow.interface';
 import { Cow } from './cow.model';
 import { IGenericResponse } from '../../../interfaces/common';
+import ApiError from '../../../errors/ApiError';
+import httpStatus from 'http-status';
+import { User } from '../user/user.model';
+import { JwtHelpers } from '../../../helpers/jwtHelpers';
+import config from '../../../config';
+import { Secret } from 'jsonwebtoken';
 
 const createCow = async (cowData: ICow): Promise<ICow> => {
   const result = (await Cow.create(cowData)).populate('seller');
@@ -72,16 +78,74 @@ const getSingleCow = async (id: string): Promise<ICow | null> => {
 
 const updateCow = async (
   id: string,
-  payload: Partial<ICow>
+  payload: Partial<ICow>,
+  token: string
 ): Promise<ICow | null> => {
+  //checking wheater the updated data is emty object or not
+  if (!Object.keys(payload).length) {
+    throw new ApiError(
+      httpStatus.BAD_REQUEST,
+      'You did not enter anything to update !'
+    );
+  }
+  //checking the cow is exist or not
+  const isExist = await Cow.findOne({ _id: id });
+  if (!isExist) {
+    throw new ApiError(httpStatus.NOT_FOUND, 'Cow not found');
+  }
+
+  const isSEllerExist = await User.findOne({ _id: payload?.seller });
+
+  if (payload?.seller) {
+    if (!isSEllerExist) {
+      throw new ApiError(
+        httpStatus.NOT_FOUND,
+        'If you want to update, you have to pay the seller id'
+      );
+    }
+  }
+
+  const verifiedUser = JwtHelpers.verifiedToken(
+    token,
+    config.jwt.secret as Secret
+  );
+
+  if (!verifiedUser) {
+    throw new ApiError(httpStatus.UNAUTHORIZED, 'You are not authorized');
+  }
+
+  const cow = await Cow.findById(id);
+
+  if (cow?.seller.toString() !== verifiedUser?.userId) {
+    throw new ApiError(httpStatus.FORBIDDEN, 'Forbidden..');
+  }
+
   const result = await Cow.findByIdAndUpdate({ _id: id }, payload, {
     new: true,
-  });
+  }).populate('seller');
 
   return result;
 };
 
-const deleteCow = async (id: string): Promise<ICow | null> => {
+const deleteCow = async (id: string, token: string): Promise<ICow | null> => {
+  const isCowExist = await Cow.findById(id);
+  if (!isCowExist) {
+    throw new ApiError(httpStatus.NOT_FOUND, 'Cow not found');
+  }
+
+  const verifiedUser = JwtHelpers.verifiedToken(
+    token,
+    config.jwt.secret as Secret
+  );
+
+  if (!verifiedUser) {
+    throw new ApiError(httpStatus.UNAUTHORIZED, 'You are not authorized');
+  }
+
+  if (isCowExist?.seller.toString() !== verifiedUser?.userId) {
+    throw new ApiError(httpStatus.FORBIDDEN, 'Forbidden..');
+  }
+
   const result = await Cow.findByIdAndDelete(id);
 
   return result;
